# api-doc.yaml
swagger: '2.0'
info:
  description: 'Webservice documentation for the LCA Collaboration Server'
  version: '1.1.1'
  title: 'LCA Collaboration Server Webservice'
  contact:
    email: 'greve@greendelta.com'
host: 'data.openlca.org'
schemes:
- https
- http
definitions:
...
  RepositoryDescriptorSearchResult:
    allOf:
    - $ref: '#/definitions/SearchResult'
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/RepositoryDescriptor'
    required:
      - resultInfo
...
  RepositoryDescriptor:
    type: object
    properties:
      group: 
        description: 'The group of the repository'
        type: string
        example: 'examples'
      name: 
        description: 'The name of the repository'
        type: string
        example: 'reference_data'
      groupIsUserNamespace:
        description: 'True if the group is a user namespace'
        type: boolean
        default: false
      settings:
        $ref: '#/definitions/RepositorySettings'
    required:
      - group
      - name
...
parameters:
  group:
    name: 'group'
    description: 'The group of the repository'
    in: path
    type: string
    required: true
  repo:
    name: 'repo'
    description: 'The name of the repository'
    in: path
    type: string
    required: true
...
  page:
    name: 'page'
    description: 'page value for pagination, if the value is 0, no paging is applied and all values will be returned'
    in: query
    type: integer
    format: int32
    default: 1
  pageSize:
    name: 'pageSize'
    description: 'Defines the page size for pagination if page is not 0'
    in: query
    type: integer
    format: int32
    default: 10
  filter:
    name: 'filter'
    description: 'Filters results that contain the phrase in their name'
    in: query
    type: string
responses:
  NotLoggedIn:
    description: 'Not authenticated'
  NoReadPerm:
    description: 'No permission to READ in {group}/{repo}'
...
  RepoNotFound:
    description: '{group}/{repo} not found'
...
  UnsupportedSchema:
    description: 'Repository json schema version is not supported'
...
paths:
...
  /ws/public/repository/{group}/{repo}:
    get:
      tags:
      - 'Repository'
      summary: 'Get public repository settings'
      description: 'Loads settings of {group}/{repo}. For this operation the client needs permission to READ in {group}/{repo}'
      operationId: 'getPublicRepository'
      produces:
      - application/json
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      responses:
        200:
          description: 'Returns the settings for {group}/{repo}'
          schema:
            $ref: '#/definitions/RepositoryDescriptor'
        403:
          $ref: '#/responses/NoReadPerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
  /ws/public/repository/avatar/{group}/{repo}:
    get:
      tags:
      - 'Repository'
      summary: 'Get repository avatar'
      description: 'Loads the avatar for {group}/{repo}. For this operation the client needs permission to READ in {group}/{repo}'
      operationId: 'getPublicRepositoryAvatar'
      produces:
      - application/octet-stream
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      responses:
        200:
          description: 'Returns the avatar for {group}/{repo} as byte array'
        403:
          $ref: '#/responses/NoReadPerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
...
  /ws/repository:
    get:
      tags:
      - 'Repository'
      summary: 'List accessible repositories'
      description: 'Loads a list of repositories that are accessible for the client'
      operationId: 'getRepositories'
      produces:
      - application/json
      parameters:
      - $ref: '#/parameters/page'
      - $ref: '#/parameters/pageSize'
      - $ref: '#/parameters/filter'
      - name: 'group'
        description: 'Filters repositories by group (exact match)'
        in: query
        type: string
      - name: 'module'
        description: 'Module specific filter. If value is "REVIEW", the list of accessible repositories is filtered for those the user can manage tasks in'
        in: query
        type: string
        enum: ['REVIEW']
      responses:
        200:
          description: 'Returns a list of accessible repositories'
          schema:
            $ref: '#/definitions/RepositoryDescriptorSearchResult'
        401:
          $ref: '#/responses/NotLoggedIn'
  /ws/repository/{group}/{repo}:
    get:
      tags:
      - 'Repository'
      summary: 'Get repository settings'
      description: 'Loads settings and permissions of {group}/{repo}. For this operation the client needs permission to READ in {group}/{repo}'
      operationId: 'getRepository'
      produces:
      - application/json
      responses:
        200:
          description: 'Returns the settings and permissions for {group}/{repo}'
          schema:
            $ref: '#/definitions/Repository'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoReadPerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
    post:
      tags:
      - 'Repository'
      summary: 'Create a new repository'
      description: 'Creates a new repository in {group} with name {repo}. For this operation the client needs permission to WRITE to {group} and general permission to create repositories'
      operationId: 'createRepository'
      produces:
      - application/json
      responses:
        201:
          description: 'Returns the repository descriptor of the newly created repository'
          schema:
            $ref: '#/definitions/RepositoryDescriptor'
        400:
          description: 'Missing input for group or name, not existing group or invalid name. Valid names consist of four characters and only contain characters, number and _. Also some specific words are reserved'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoWritePermGroup'
        406:
          $ref: '#/responses/UnsupportedSchema'
        409:
          description: '{repo} already exists in {group}'
    delete:
      tags:
      - 'Repository'
      summary: 'Delete a repository'
      description: 'Deletes {group}/{repo}. For this operation the client needs permission to DELETE {group}/{repo}'
      operationId: 'deleteRepository'
      responses:
        200:
          description: '{group}/{name} was successfully deleted'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoDeletePerm'
        406:
          $ref: '#/responses/UnsupportedSchema'
    parameters:
    - $ref: '#/parameters/group'
    - $ref: '#/parameters/repo'
  /ws/repository/avatar/{group}/{repo}:
    get:
      tags:
      - 'Repository'
      summary: 'Get repository avatar'
      description: 'Loads the avatar for {group}/{repo}. For this operation the client needs permission to READ in {group}/{repo}'
      operationId: 'getRepositoryAvatar'
      produces:
      - application/octet-stream
      responses:
        200:
          description: 'Returns the avatar for {group}/{repo} as byte array'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoReadPerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
    put:
      tags:
      - 'Repository'
      summary: 'Set repository avatar'
      description: 'Set the avatar for {group}/{repo}. For this operation the client needs permission to WRITE in {group}/{repo}'
      operationId: 'setRepositoryAvatar'
      consumes:
      - multipart/form-data
      produces:
      - application/octet-stream
      responses:
        200:
          description: 'Returns the avatar for {group}/{repo} as byte array'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoWritePerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
    parameters:
    - $ref: '#/parameters/group'
    - $ref: '#/parameters/repo'
  /ws/repository/meta/{group}/{repo}:
    get:
      tags:
      - 'Repository'
      summary: 'Get repository meta information'
      description: 'Loads meta information for {group}/{repo}. For this operation the client needs permission to READ in {group}/{repo}'
      operationId: 'getRepositoryMetaInfo'
      produces:
      - application/json
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      responses:
        200:
          description: 'Returns repository meta information'
          schema:
            $ref: '#/definitions/RepositoryMetaInfo'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoReadPerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
  /ws/repository/export/{group}/{repo}:
    get:
      tags:
      - 'Repository'
      summary: 'Export a repository (migrate)'
      description: 'Exports {group}/{repo} for migration to another collaboration server (only data and repo info is exported, comments and membership information are dropped). For this operation the client needs permission to READ in {group}/{repo}'
      operationId: 'exportRepository'
      produces:
      - application/octet-stream
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      responses:
        200:
          description: 'Returns a zip file containing the complete repository data directory, containing the data sets, commit history, settings and meta information'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoReadPerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
  /ws/repository/import/{group}/{repo}:
    post:
      tags:
      - 'Repository'
      summary: 'Import data into an empty repository'
      description: 'Imports either an exported repository from another collaboration server or a JSON-LD zip file exported from openLCA. If {group}/{repo} is not empty, existing data will be overwritten. For this operation the client needs permission to WRITE in {group}/{repo}'
      operationId: 'importRepository'
      consumes:
      - multipart/form-data
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - name: 'commitMessage'
        description: 'The commit message in case of format=json-ld'
        in: formData
        type: string
      - name: 'file'
        description: 'The file to be imported'
        in: formData
        type: string
      - name: 'format'
        description: 'Either json-ld or empty (empty=import exported repository)'
        in: query
        type: string
      responses:
        200:
          description: 'The data was successfully imported'
        400:
          description: 'Missing input for commit message, when format=json-ld'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoWritePerm'
        406:
          $ref: '#/responses/UnsupportedSchema'
  /ws/repository/move/{group}/{repo}/{newGroup}/{newRepo}:
    post:
      tags:
      - 'Repository'
      summary: 'Move an existing repository'
      description: 'Move {group}/{repo} to {newGroup}/{newRepo} or rename it within the same group if {group}={newGroup}. For this operation the client needs permission to WRITE in {newGroup} and permission to READ in {group}/{repo}'
      operationId: 'moveRepository'
      produces:
      - application/json
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - name: 'newGroup'
        description: 'The name of the group to move {group}/{repo} to (if it equals the existing group, the repository is renamed)'
        in: path
        type: string
        required: true
      - name: 'newRepo'
        description: 'The new name of the repo'
        in: path
        type: string
        required: true
      responses:
        200:
          description: '{group}/{repo} was successfully moved/renamed'
        400:
          description: '{newGroup} does not exist'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoWritePerm'
        406:
          $ref: '#/responses/UnsupportedSchema'
        409:
          description: '{newGroup}/{newRepo} already exists'
  /ws/repository/clone/{group}/{repo}/{commitId}/{newGroup}/{newRepo}:
    post:
      tags:
      - 'Repository'
      summary: 'Clone an existing repository'
      description: 'Clone an existing repository to {newGroup}. For this operation the client needs permission to WRITE in {newGroup} and permission to READ in {group}/{repo}'
      operationId: 'cloneRepository'
      produces:
      - application/json
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - name: 'commitId'
        description: 'If specified only data until {commitId} are cloned'
        in: path
        type: string
        required: true
      - name: 'newGroup'
        description: 'The name of the group to move the repository to (if it equals the existing group, the repository is renamed'
        in: path
        type: string
        required: true
      - name: 'newRepo'
        description: 'The new name of the repo'
        in: path
        type: string
        required: true
      responses:
        200:
          description: 'The repository was successfully moved/renamed'
        400:
          description: '{newGroup} does not exist'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoWritePerm'
        406:
          $ref: '#/responses/UnsupportedSchema'
        409:
          description: '{newGroup}/{newRepo} already exists'
...
  /ws/membership/{group}/user/{username}/{role}:
    post:
      tags:
      - 'Membership (Group)'
      summary: 'Add a user to a group'
      description: 'Add {username} to {group} with {role}. For this operation the client needs permission to CHANGE_ROLE in {group}'
      operationId: 'addUserMembershipForGroup'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        201:
          description: '{username} was added to {group} with {role}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePermGroup'
        404:
          $ref: '#/responses/GroupNotFound'
        409:
          description: '{username} already is a direct member of {group}'
    put:
      tags:
      - 'Membership (Group)'
      summary: 'Update a users role in a group'
      description: 'Update the role of {username} in {group} to {role}. For this operation the client needs permission to CHANGE_ROLE in {group}'
      operationId: 'updateUserMembershipForGroup'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        200:
          description: 'The role of {username} was changed to {role} in {group}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePermGroup'
        404:
          description: '{group} not found or {username} is not a member'
  /ws/membership/{group}/user/{username}:
    delete:
      tags:
      - 'Membership (Group)'
      summary: 'Delete a user from a group'
      description: 'Deletes {username} from {group}. For this operation the client needs permission to CHANGE_ROLE in {group}'
      operationId: 'deleteUserMembershipForGroup'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/username'
      responses:
        200:
          description: '{username} was deletes from {group}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePermGroup'
        404:
          description: '{group} not found or {username} is not a member'
...
  /ws/membership/{group}/{repo}/user/{username}/{role}:
    post:
      tags:
      - 'Membership (Repository)'
      summary: 'Add a user to a repository'
      description: 'Add {username} to {group}/{repo} with {role}. For this operation the client needs permission to CHANGE_ROLE in {group}/{repo}'
      operationId: 'addUserMembershipForRepository'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        201:
          description: '{username} was added to {group}/{repo} with {role}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePerm'
        404:
          $ref: '#/responses/RepoNotFound'
        406:
          $ref: '#/responses/UnsupportedSchema'
        409:
          description: '{username} already is a direct member of {group}/{repo}'
    put:
      tags:
      - 'Membership (Repository)'
      summary: 'Update a users role in a repository'
      description: 'Update the role of {username} in {group}/{repo} to {role}. For this operation the client needs permission to CHANGE_ROLE in {group}/{repo}'
      operationId: 'updateUserMembershipForRepository'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - $ref: '#/parameters/username'
      - $ref: '#/parameters/role'
      responses:
        200:
          description: 'The role of {username} was changed to {role} in {group}/{repo}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePerm'
        404:
          description: '{group}/{repo} not found or {username} is not a member'
        406:
          $ref: '#/responses/UnsupportedSchema'
  /ws/membership/{group}/{repo}/user/{username}:
    delete:
      tags:
      - 'Membership (Repository)'
      summary: 'Delete a user from a repository'
      description: 'Deletes {username} from {group}/{repo}. For this operation the client needs permission to CHANGE_ROLE in {group}/{repo}'
      operationId: 'deleteUserMembershipForRepository'
      parameters:
      - $ref: '#/parameters/group'
      - $ref: '#/parameters/repo'
      - $ref: '#/parameters/username'
      responses:
        200:
          description: '{username} was deletes from {group}/{repo}'
        401:
          $ref: '#/responses/NotLoggedIn'
        403:
          $ref: '#/responses/NoChangeRolePerm'
        404:
          description: '{group}/{repo} not found or {username} is not a member'
        406:
          $ref: '#/responses/UnsupportedSchema'
...
  /ws/user:
    get:
      tags: 
      - 'User'
      summary: 'Get users'
      description: 'Loads a list of user descriptors'
      operationId: 'getUsers'
      parameters:
      - $ref: '#/parameters/page'
      - $ref: '#/parameters/pageSize'
      - $ref: '#/parameters/filter'
      - name: 'module'
        description: 'Load a module specific list, MESSAGING loads all users that can be messaged, REVIEW loads all users that have the permission to REVIEW in {repositoryPath}'
        type: string
        enum: ['MESSAGING', 'REVIEW']
        in: query
      - name: 'repositoryPath'
        description: 'The repository id to filter users from (only applies if module=REVIEW)'
        type: string
        in: query
      produces:
      - application/json
      responses:
        200:
          description: 'A list of user descriptors'
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        401:
          $ref: '#/responses/NotLoggedIn'
        404:
          $ref: '#/responses/UserNotFound'
...